+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 CHANGELOG
 =========
 
+1.14.0-dev
+==========
+
+* feature: add support for sagemaker-tensorflow-serving container
+* feature: Estimator: make input channels optional
+
 1.13.0
 ======